Rename prototype gcds-site-menu
