--- packages/web/src/components/gcds-site-menu/i18n/i18n.js
@@ -1,18 +1,0 @@
-const I18N = {
-    "en": {
-        navLabel: "Main navigation - Use the enter key to select a menu item and travel to its page. Use the left and right arrow keys to navigate between menu and submenu items. Use the right arrow key to open submenus when they are available. Use the left arrow or escape keys to close a menu.",
-        mobileTriggerClose: "Close",
-        mobileTriggerLabel: "Use the enter key to open or close the main menu.",
-        submenuButtonText: "Open or close {$t}'s submenu.",
-        back: "Back"
-    },
-    "fr": {
-        navLabel: "Navigation principale - Utiliser la touche d'entrée pour sélectionner un élément du menu et voyager à la page indiquée. Utiliser les flèches gauches et droites pour naviguer entre les éléments et les sous-éléments du menu. Ouvrir les sous-éléments du menu avec la flèche droite lorsqu'il sont disponible. Fermer le menu avec la flèche gauche ou la touche d'échappement.",
-        mobileTriggerClose: "Fermer",
-        mobileTriggerLabel: "Utiliser la touche Entrée pour ouvrir ou fermer le menu principal.",
-        submenuButtonText: "Ouvrir ou fermer l'élément du sous-menu {$t}.",
-        back: "Retour"
-    }
-}
-
-export default I18N;
