+++ packages/web/src/components/gcds-site-menu/readme.md
@@ -1,4 +1,4 @@
-# gcds-site-menu
+# gcds-site-menu1
 
 
 
@@ -7,27 +7,12 @@
 
 ## Properties
 
-| Property                     | Attribute        | Description            | Type                                       | Default     |
-| ---------------------------- | ---------------- | ---------------------- | ------------------------------------------ | ----------- |
-| `alignment`                  | `alignment`      | Menu alignment         | `"center" \| "left" \| "right" \| "split"` | `'left'`    |
-| `desktopLayout` _(required)_ | `desktop-layout` | Desktop layout         | `"sidebar" \| "topbar"`                    | `undefined` |
-| `mobileLayout` _(required)_  | `mobile-layout`  | Mobile layout          | `"drawer"`                                 | `undefined` |
-| `position`                   | `position`       | Sticky navigation flag | `"static" \| "sticky"`                     | `'static'`  |
+| Property    | Attribute   | Description                   | Type                                       | Default     |
+| ----------- | ----------- | ----------------------------- | ------------------------------------------ | ----------- |
+| `alignment` | `alignment` | Menu alignment                | `"center" \| "left" \| "right" \| "split"` | `'left'`    |
+| `label`     | `label`     | Label for navigation landmark | `string`                                   | `undefined` |
+| `position`  | `position`  | Sticky navigation flag        | `"static" \| "sticky"`                     | `'static'`  |
 
-
-## Dependencies
-
-### Depends on
-
-- [gcds-button](../gcds-button)
-
-### Graph
-```mermaid
-graph TD;
-  gcds-site-menu --> gcds-button
-  gcds-button --> gcds-icon
-  style gcds-site-menu fill:#f9f,stroke:#333,stroke-width:4px
-```
 
 ----------------------------------------------
