--- packages/web/src/components/gcds-site-menu1/readme.md
@@ -1,19 +1,0 @@
-# gcds-site-menu1
-
-
-
-<!-- Auto Generated Below -->
-
-
-## Properties
-
-| Property    | Attribute   | Description                   | Type                                       | Default     |
-| ----------- | ----------- | ----------------------------- | ------------------------------------------ | ----------- |
-| `alignment` | `alignment` | Menu alignment                | `"center" \| "left" \| "right" \| "split"` | `'left'`    |
-| `label`     | `label`     | Label for navigation landmark | `string`                                   | `undefined` |
-| `position`  | `position`  | Sticky navigation flag        | `"static" \| "sticky"`                     | `'static'`  |
-
-
-----------------------------------------------
-
-*Built with [StencilJS](https://stenciljs.com/)*
